+++ src/syntax_placeholders.jl
@@ -1,7 +1,4 @@
-## Syntax data.
-
 ## -----------------------------------------------------------------------------------------
-
 ## Syntax placeholders
 
 """
